Add a flag to disable sf prefer high cap
This change investigates disabling SF prefer_high_cap,
by adding a flag that disables the behavior regardless of what the
powerhint config says.

Since the powerhint behavior was added in a special section of the
json and is not linked with an action, it cannot be flagged directly
in the powerhint file, so the associated code is flagged instead.

Test: manual
Bug: 434909956
Flag: powerhal.flags.ramp_down_sf_prefer_high_cap
Change-Id: Ide0bf7763dd0ccc6df924e95cea9893f3a499453